requirements.txt
fastapi
uvicorn[standard]
pillow
python-dotenv
bakong_khqr
